refactor: Networking Error 그룹으로 분리
