+++ WeatherForecast/WeatherForecast/Infrastructure/Networking/Error/HTTPResponseError.swift
@@ -1,0 +1,12 @@
+//
+//  HTTPResponseError.swift
+//  WeatherForecast
+//
+//  Created by Jason on 2023/03/28.
+//
+
+import Foundation
+
+enum HTTPResponseError: Error {
+    case error(statusCode: Int, description: String)
+}
