+++ WeatherForecast/WeatherForecast/Infrastructure/Networking/Error/NetworkEntityLoadingError.swift
@@ -1,0 +1,13 @@
+//
+//  NetworkEntityLoadingError.swift
+//  WeatherForecast
+//
+//  Created by Jason on 2023/03/28.
+//
+
+import Foundation
+
+enum NetworkEntityLoadingError: Error {
+    case networkFailure
+    case invalidData
+}
