+++ WeatherForecast/WeatherForecast/Infrastructure/Networking/Error/NetworkError.swift
@@ -1,0 +1,12 @@
+//
+//  NetworkError.swift
+//  WeatherForecast
+//
+//  Created by Jason on 2023/03/28.
+//
+
+import Foundation
+
+enum NetworkError: Error {
+    case notConnected
+}
